Adjust escaping and globbing to work with env vars passing
Now that we're passing in values through environment variables, we don't
need the extra layer of escaping that was present in the code because
the shell script was getting templated by GH actions.

Also, since this layer is now gone, we stop the shell from globbing in
the `check` step to avoid globbing the find arguments.